changing BulkResponseItem.Error to json.rawMessage
With ES 2.3, when BulkResponseItem.Error comes back ES provides a json object
rather than a string.  This causes the json.Unmarshal method to fail and the
application exits.  Treating the response as a json.RawMessage allows
json.Unmarshal to treat all error responses as a string regardless of json
data type.